requirements-lang.txt
langchain[all]==0.1.8
pyodbc==5.1.0
python-dotenv==1.0.1
openai==1.12.0
langchain_openai==0.0.6
SQLAlchemy==2.0.27
streamlit==1.31.1
asyncio==3.4.3
